fix bugs in read_links() and read_demand()
- if an attribute is None, we need to set it up using the default
  value from the default constructor. Passng None to __init() will
  lead to None to be assigned
- if volume of an record is None, discard it
- fix the potential bug that _get_interval_id() returns a float such as
  9.0 which will ruin the following process. Ensure the return is int

+++ path4gmns/accessibility.py
@@ -21,7 +21,7 @@ def _get_interval_id(t):
         return 0
 
     if (t % (MIN_TIME_BUDGET+BUDGET_TIME_INTVL)) == 0:
-        return (t/(MIN_TIME_BUDGET+BUDGET_TIME_INTVL))
+        return int(t/(MIN_TIME_BUDGET+BUDGET_TIME_INTVL))
 
     return int(t/(MIN_TIME_BUDGET+BUDGET_TIME_INTVL)) + 1
